Split torque boundaries from regression tests
